django-rest/mm_blog/requirements.txt
amqp==5.2.0
asgiref==3.8.1
astroid==3.2.4
billiard==4.2.0
celery==5.4.0
cffi==1.17.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.0
dill==0.3.8
Django==5.1
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.15.2
iniconfig==2.0.0
isort==5.13.2
kombu==5.4.0
mccabe==0.7.0
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
pycparser==2.22
pytest==8.3.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
tomlkit==0.13.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
